Change variable to boolean in companion.

+++ companion/src/firmwares/edgetx/yaml_logicalswitchdata.cpp
@@ -116,8 +116,8 @@ Node convert<LogicalSwitchData>::encode(const LogicalSwitchData& rhs)
   node["delay"] = rhs.delay;
   node["duration"] = rhs.duration;
   node["andsw"] = YamlRawSwitchEncode(RawSwitch(rhs.andsw));
-  node["lsPersist"] = rhs.lsPersist;
-  node["lsState"] = rhs.lsState;
+  node["lsPersist"] = (int)rhs.lsPersist;
+  node["lsState"] = (int)rhs.lsState;
 
   return node;
 }
